pyproject.toml
[project]
name = "llm-structured-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datamodel-code-generator>=0.33.0",
    "genson>=1.3.0",
    "langchain-aws>=0.2.33",
    "langchain-core>=0.3.76",
    "pdfplumber>=0.11.7",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "streamlit-code-editor>=0.1.22",
    "streamlit[pdf]>=1.49.1",
]
